+++ app/commands/sources.py
@@ -14,5 +14,6 @@ def sources_command(bot, update, chat_info):
 https://bitfinex.com - 15min (API limits😭)
 https://bittrex.com - 1min
 [https://bx.in.th](https://bx.in.th/ref/s9c3HU/) - 1min
+https://sp-today.com - 60min
 https://fixer.io - 60min
 https://openexchangerates.org - 60min''')