+++ app/settings.py
@@ -72,6 +72,12 @@ CELERYBEAT_SCHEDULE = {
         'args': ('app.exchanges.FixerExchange',),
         'options': {'time_limit': 60, 'once': {'timeout': 60}}
     },
+    'exchange_updater_SpTodayExchange': {
+        'task': 'app.tasks.exchange_updater',
+        'schedule': crontab(minute=1, hour='*/1'),
+        'args': ('app.exchanges.SpTodayExchange',),
+        'options': {'time_limit': 60, 'once': {'timeout': 60}}
+    },
     'delete_expired_rates': {
         'task': 'app.tasks.delete_expired_rates',
         'schedule': crontab(minute=5, hour='*/1'),