+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 ##### 1.1
 * added BTT - BitConnect currency
-* added [bx.in.th](https://bx.in.th/ref/s9c3HU/) exchange
+* added [bx.in.th](https://bx.in.th/ref/s9c3HU/) exchange for THB pairs
+* added sp-today.com exchange for SYP pairs
 * /settings command, only for private chats: language, default currency settings
 * fixes and improvements